Add 'newTagRevision' property to fix tagging problem
Fixes #46
Using 'newTagRevision' property eliminates problem with wrong tagging, when release is tagged with revision reference that the release doesn't come from, but is the latest one on master branch.
To solve the problem revision from which release is created has to be specified with GH API 'target_commitish' parameter. To get the revision default GitHub Actions environmental variable 'GITHUB_SHA' is used.