Simplify token configuration
Currently 2 different tokens are used to authenticate in GitHub API: 'readOnlyToken' that is used for fetching tickets and 'writeToken' to post releases.
Read token can be checked-in into the *.gradle file but this can generate token overuse that counts against the API rate limit.
Using different types of tokens complicates also plugin configuration.
This commit adds 'githubToken' property that is used both for reading and posting and deprecates 'readOnlyToken' and 'writeToken'. It should be supplied by environmental variable that is configured in CI system.